requirements.txt
python-dotenv>=1.0.0
openai>=1.0.0
requests>=2.28.0
pypdf>=3.0.0
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
numpy>=1.24
pandas>=2.0
tqdm>=4.65
rich>=12.6.0
pytest>=7.0
uvicorn>=0.22.0
gradio>=3.38.0
beautifulsoup4>=4.12.2
pyarrow>=11.0.0
fastparquet>=0.9.0